refactor)Board): move out from PVZ.h
