+++ pvzclass/Classes/Griditem.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include "../PVZ.h"
+#include "Board.hpp"
 
 namespace PVZ
 {
