+++ pvzclass/pvzclass.vcxproj.filters
@@ -860,5 +860,8 @@
     <ClInclude Include="Classes\GameObject.hpp">
       <Filter>头文件\Classes</Filter>
     </ClInclude>
+    <ClInclude Include="Classes\Board.hpp">
+      <Filter>头文件\Classes</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
